[WHM-11] Add BottomNavigationView - Refactoring
