+++ composeApp/src/desktopMain/kotlin/com/yttydev/whm/main.kt
@@ -2,6 +2,7 @@ package com.yttydev.whm
 
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import com.yttydev.whm.presentation.base.App
 
 fun main() = application {
     Window(
