+++ composeApp/src/iosMain/kotlin/com/yttydev/whm/MainViewController.kt
@@ -1,5 +1,6 @@
 package com.yttydev.whm
 
 import androidx.compose.ui.window.ComposeUIViewController
+import com.yttydev.whm.presentation.base.App
 
 fun MainViewController() = ComposeUIViewController { App() }
