+++ composeApp/src/wasmJsMain/kotlin/com/yttydev/whm/main.kt
@@ -2,6 +2,7 @@ package com.yttydev.whm
 
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
+import com.yttydev.whm.presentation.base.App
 import kotlinx.browser.document
 
 @OptIn(ExperimentalComposeUiApi::class)
